pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-werewolf"
version = "0.1.0"
description = "An AI-powered Werewolf game where LLM agents play against each other"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["ai", "llm", "game", "werewolf", "multi-agent"]
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
werewolf = "src.main:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
